Add preview on codelens
